release: Implement PT label checking
From now on, the stories labeled 'point me' will be listed on release start and
the user will be asked to confirm that he or she really wants to continue.

On release stage, certail labels cause the story check to be skipped, notable it
is always 'dupe' and 'wontfix', but more labels can be added in the project
configuration file, under pivotaltracker.labels.skip_release_check, which is a
list of labels to cause the check skip behaviour.