Refactor the repository structure
Drop the part of the code base written in Bash, it is no longer needed.
The git-trunk directory becomes the top level directory, making the whole thing
nicely go-gettable as github.com/salsita/salsaflow.

Also renamed bin/hooks/commit-msg to bin/hooks/salsaflow-commit-msg so that it
can be installed as salsaflow-commit-msg executable directly.

Change-Id: 808d84ce8f