Makefile: Update to reflect the new paths, cleanup
The system-wide install target was actually not that useful, so the install
target that was left in the Makefile just installs the executables into the
project workspace.

The update-deps target was updates to include all 3rd-party packages that are
being used right now. It is handly for development purposes since just running
go get ./... modifies the project repository, which is not always desired.

Change-Id: 380fff8dc8