story start: Polish the code and fix minor issues
Aligned the output to be more consistent with other commands. The list of
available stories was refactored, but also some RUN messages were rewritten to
represent the task name, i.e. to say "Fetch stories" instead of "Fetching
stories". This is how the tasks are names elsewhere, so let's stay consistent.

Also fixed a few minor issues, for example that the new story branch was
actually not being checked out.

Change-Id: fb3fce7207