story start: Fetch the remote repository
The remote repository was not being updated during story start, which makes the
story branch check weaker by possibly missing the remote branches. There is no
reason not to update the repository since we anyway need to access remote
services and go online to download stories from the issue tracker.

Change-Id: 79d76d3cb0