release start: Review the command and fix errors
Update release start command to the latest conventions and also fix some
critical bugs found on the way, particularly somehow completely wrong order of
steps being executed. For example, the version string was first read from the
trunk branch, then the repository was updated from the remote. Exactly the
opposite to what it should be.

Change-Id: 2ea30ab39f