repo init: Invoke automagically
From now on the user does not have to call repo init manually, although the
option is still there. In app.Init, it is checked whether the repo is
initialised, and when it is not, repo init is invoked automatically. So the
repository is initialised when the user runs the first SalsaFlow command.

Change-Id: 81f01eab05
Story-Id: SF-26