Implement new commands: pkg install, pkg upgrade
pkg install allows the user to install particular version of SalsaFlow. The
executables are downloaded from the relevant GitHub release and the current
executables are replaced.

pkg upgrade works in a similar way, it just automatically finds the most recent
release and upgrades the executables in case the detected release is more recent
than what is installed.

Change-Id: a3866677ed
Story-Id: SF-25