hooks: Implement review post process changes
The commit-msg hook no longer needs to check for Story-Id. This is what review
post command is doing from now on. The only thing necessary is to insert
Change-Id.

The pre-push hook shall reject any commits not complying with the SalsaFlow
requirements, i.e. the commits not having either Change-Id or Story-Id.

Change-Id: e43075869e
Story-Id: SF-15