IssueTracker: Add StoriesInProgress method
The new method lists the stories that are currently in progress, i.e. the
develop may be still actively working on them.

This is being implemented so that the new review post command can download the
stories in progress and ask the user to choose the story that the commit to be
posted is going to be associated with.

Change-Id: 4569e88140
Story-Id: SF-15