review post: Use new variable to keep commit msg
When amending the commit message during review post, the new message is saved in
the commit object that is being processes. But that is not the commit that is
going to be created with the new commit message. It makes more sense to keep the
value in a new variable, not changing the commit struct at all.

Story-Id: SF-15
Change-Id: 7240d5b729