hooks/pre-push: Check only the core branches
It was getting really annoying the the pre-push hook always required the
Story-Id tag to be present in the commit message, even though the push was a WIP
push and was not updating any branch important for SalsaFlow.

This patch is modifying the hook to only check updates to the core branches,
i.e. trunk, release, client and master.

Change-Id: 4aaa48ef34
Story-Id: SF-15