modules/jira: Display only coding-related issues
This patch limits all the JQL search queries to just return the issues
that are of one of the coding-related types. This ensures that the user
does not choose an issue that is not related to coding at all and it
also makes the list smaller, which is always good.

Story-Id: SF-62
Change-Id: 02f47e086b