config: Refactor config handling heavily
Basically the whole config handling was rewritten and refactored to make more
sense for the current code base. There is no global config loading function that
must be called at the beginning of every command any more, the config package
takes care only of reading and parsing the config files.

The rest of the config processing was pushed into the packages where it makes
the most sense, i.e. the git branch names can be got by calling git.LoadConfig.
Modules and their config works in the same way. This makes the whole config work
on demand, getting rid of initialisation problems. If you need config, simply
call the relevant function to get the right config struct.

Also the local config file is now read from the working tree rather than from a
branch. This makes it possible to actually make the trunk branch name
configurable.

In general more things are configurable now, including the project remote, which
is by default origin, but can be set with git config as salsaflow.remote.

Change-Id: 5f57ef0f29
Story-Id: SF-67