version: Make operations configurable
From now on it is not needed to just use package.json for all projects
that want to use SalsaFlow. Actually, custom scripts to handle version
string manipulation must be specified in the local config file.

The GET operation is configurable through scripts.get_version key in the
local config file. The value must be a relative path from the repository
root. The script found on this path is executed in the repository root
and it is expected to print the version string to stdout.

The SET operation is configurable through scripts.set_version key in the
local config file. The value must be a relative path from the repository
root. The script found on this path is executed in the repository root
and it receives the version string to be set as the first argument. It
is then expected to write the version string to all the files where it
is needed.

The scripts can, naturally, print errors to stderr and return a non-zero
exit status to signal a failure.

Change-Id: 956fd3f744
Story-Id: SF-63