release stage: Ensure no commits are left behind
From now on, release stage checks to make sure that all changes assigned
to the stories that are associated with the release to be stages are
actually present on the release branch.

In other words, it warns the user in case there are some changes being
left behind, i.e. they have not been cherry-picked onto the release
branch yet. The user is prompted to confirm the staging process if that
is the case.

Change-Id: b7f985358e
Story-Id: SF-84