jira: Fix staging in RunningRelease implementation
runninRelease.EnsureStageable was not making sure that also the
sub-tasks are ok, i.e. they are in the Done state.

runningRelease.Stage was trying to stage all issues assigned to the
current release. This should be limited to the issues that are Tested,
otherwise the transition does not make sense.

The checks are only checking states, not issue types. This is to make
the implementation a bit more robust against changes in the workflow, or
rather adding new issue types.

Change-Id: fe6a6a4e31
Story-Id: SF-84