release stage: Actually release the stories
release stage is now actually releasing the affected stories, which is
what was forgotten in the original implementation of release stage.

The command was also a bit refactored and improved. The push step is not
the last step any more, the story release step is the last step now. In
case there is an error in the issue tracker release step, the user is
asked to fix the inconsistencies manually. This is easier than rolling
back after failing to push. And the API calls should mostly work just
fine, all conditions are checked before the push, so the last them is
just about performing the API call to modify the stories.

Change-Id: e61190c344
Story-Id: SF-78