config: Change local config structure
To make it possible to bootstrap a repository, the custom scripts paths
are from now defined by SalsaFlow, relative to the repository root. This
also means that the relevant local configuration section is no longer
needed.

The new local configuration structure is:

  .salsaflow/
    config.yml
    scripts/
      get_version
      set_version
      ...

Also to make it possible to write scripts that would work across
platforms, a platform suffix can be attached to the script file. For
example, set_version_windows is being used on Windows, set_version_unix
on other operating systems. Naturally, if no platform suffix is
supplied, the same script file is used on all platforms.

Change-Id: 521fca8971
Story-Id: SF-99