Add a new command: repo bootstrap
This command can be used to generate initial configuration file in the
current repository so that the project can get up to speed with
SalsaFlow quickly.

This basically means that the local configuration file is generated
based on the flags specified. The user then has to check the files and
fill in the missing information.

It is also possible now to use so-called skeletons, which are userful
particularly for custom scripts. Telling repo bootstrap to use
'salsaflow/skeleton-golang' will make SalsaFlow clone the given
repository from GitHub and it will pour the content of the default
branch into .salsaflow directory of the repository. For now it only
makes sense to have the scripts directory there and use the given
skeleton to share these scripts across the same project types.

Change-Id: 05dd968d1a
Story-Id: SF-99