Fix release/story changes
These commands were buggy and mostly not working properly due to a few major
flaws in the implementation coming from not understanding git enough.

This patch makes sure the changes commands work properly. The patch is rather
scattered across the code base because the fixes and changes just implied
further changes in the code base. The end result should be, though, that the
release changes and story changes commands work properly now.

Change-Id: 048b5cc7f2
Story-Id: SF-88