scripts: Make it possible to go cross-platform
To make it actully possible to run distinct scripts for various platforms, a
notion of script runners was introduced. When certain script is about to be run,
the right script file is found based on the current platform and file extension.
The file extension is then used to get a runner for that script.

So, to run 'get_version' on Mac OS X, it is necessary to have a script names
'get_version_{unix|darwin}.<runner_supported_on_Mac_OS_X>', e.g
'get_version_unix.bash' or 'get_version_unix.js'.

Supported runners:
  Bash (.bash) (Unix)
  Node.js (.js) (Unix, Windows)
  cmd.exe (.bat) (Windows)
  PowerShell.exe (.ps1) (Windows)

Change-Id: 9eb8b60b91
Story-Id: SF-99