pivotaltracker: Make a new friend, Review Board
Make sure review post works well with Pivotal Tracker. This required a
new method for the common.Story interface, Tag(). This method is being
used to generate the right Story-Id tag value for the story. This is
because we actually need a special Story-Id tag value for Pivotal
Tracker so that not only the PT story ID, but also the project ID is
stored in the tag and the value can be easily used for the RB bugs
closed field.

Change-Id: beb3ee023a
Story-Id: SF-106