git: Rename StoryIds to StoryIdTags and fix stuff
The problem is that with some issue trackers, the Story-Id tag does not
necessarily hold only the story ID itself, so there is some additional parsing
required to get the story ID out of the Story-Id tag. Hence the renaming.

Also fixed the Pivotal Tracker module to work properly with the Story-Id tags
that the module is expecting.

Change-Id: a2602e7847
Story-Id: SF-108