review post: Ask the user what to do next
When the -parent flag is used, the user usually wants to do something with the
story branch once the review requests are created, checked and published. People
have been, however, forgetting to push the story branch, merge into trunk etc.
So, this patch makes review post prompt the user after the reviews are created
to know what to do next. There is a simple 3-step dialog asking the user what to
do with the modified branches.

Change-Id: 145cc926d9
Story-Id: SF-115