changes: Fix DumpStoryChanges
The issue was that the Story-Id tag is not always the story ID only, so
there is some parsing necessary to make DumpStoryChanges print the
actual story ID in the listing.

Also renamed git.Commit.ChangeId and StoryId to ChangeIdTag and
StoryIdTag respectively to make it clear that these fields do not
contain the story ID but rather the Story-Id tag.

Change-Id: 7fc17bb476
Story-Id: SF-119