story changes: Make the command more flexible
Until now it was only possible to specify the exact Story-Id tag value
to select the changes matching the tag. Now it is possible to use a
regexp to select the changes to be included in the list. This enables
the user to use only the minimum to select the right changes, or to
perhaps include multiple stories in the output.

Change-Id: dfdb4cfa2e
Story-Id: SF-120