review post: Add the option to select 'unassigned'
To be able to mark a commit as being unassigned, i.e. not assigned to
any story, we use 'unassigned' Story-Id tag value. It is, however, not
possible to choose this value in the Story-Id tag dialog.

This patch adds the option to select this value as a valid Story-Id tag
value.

Change-Id: 78cd48cc3b
Story-Id: SF-123