review post: Fix broken parent hash parsing
In case the story branch base is a merge commit, there are multiple
parents. The current code is however not parsing the output line
properly in this case since it is not splitting on whitespaces.

This patch fixes this behavior, splitting the parents hashes line and
taking the first field.

Story-Id: SF-137
Change-Id: 8ac3acbc11