release start: Change Story-Id tag parsing
In case the Story-Id tag is unassigned and it cannot be parsed using the
current issue tracker, an error is returned.

This has been changed. In case the Story-Id tag cannot be parsed using
the current issue tracker, the commit is simply skipped. This makes it
possible to use arbitrary string instead od just using 'unassigned'.

Change-Id: 8733353b4e
Story-Id: SF-140