hooks: pre-push: Make sure refs are up to date
Right now it is possible for the pre-push hook to fail saying that the
commit range it is checking is invalid. This is because the ref contains
the remote commit hash which does not exist locally in case the
reference is not up to date.

This patch implements a check that makes sure the commit exists and
explains to the user what to do in case it doesn't.

Change-Id: f5c9bf6431
Story-Id: SF-141