review post: Post RRs sequentially
Review requests are posted one by one unless -no_seq is specified. The
behaviour is changed to make sure the review requests are listed in RB
in the same order as the commits are visible in the git graph. When the
requests are fired all at once, the order in RB is undefined and it's
less obvious to the developers and reviewers what to review/fix first.

Change-Id: a5a82d1ba2
Story-Id: SF-135