release deploy: Check deployed tag format
In case the format was not the right one, i.e. vX.Y.Z, SF ended up
panicking since the deployed tag was not included in the list of release
tags associated with the given repository.

From now on the tag on the production branch is checked to make sure the
format is right, i.e. vX.Y.Z.

Story-Id: just-fixing-quickly
Change-Id: a065fda874