pre-push hook: Check commits based on timestamp
Since it is notoriously problematic to decide when to start checking
commits in the pre-push hook, a new key has been added into the local
configuration file. It is called "salsaflow_enabled_timestamp" and it
is set to the time when `repo bootstrap` was executed for the given
repository. The pre-push hook then uses this timestamp to only check
commits that happened after.

In case the timestamp is absent from the config file, the hook falls
back to the original mechanism.

Change-Id: 4b61ad22c7
Story-Id: SF-143