pkg upgrade: Compare the right GH release
There was a major bug in the release selection logic. Actually the
oldest release was selected to compare against the current one, so SF
was telling people that everything is up to date while it was not.

This should be fixed now and SF should proceed to upgrade in case a
newer release is available.

Story-Id: unassigned
Change-Id: bd464d7306