Add new code review module: github
From now on it's possible to use GitHub to do code reviews.

For every unassigned commit there is a new issue created that contains a link to
the commit to be reviewed.

For every assigned commit either a new issue is created or an existing issue is
amended in a way so that it contains the link to the new commit to be reviewed.
The operations are done in bulk, i.e. there is only one review comment added
once all commits are added for the given story (GitHub issue).

All review issues are labeled 'review'.

Change-Id: b4d5951f5a
Story-Id: SF-147