github/codereview: Change review issue body
There were some changes necessary to make it possible and not horribly ugly for
the server-side daemon to process GitHub events and pair them with the right
issues.

The change is rather ugly since it adds IssueTrackerName to common.Story just to
make it possible to insert this string into GitHub issue body. Refactoring is
planned for sure, but we need to get this working.

Change-Id: eebcaf0082
Story-Id: SF-147