hooks/pre-push: Prompt user instead of error
The pre-push hook now lists the commits missing the Story-Id tag and
prompts the user to confirm the push instead of just rejecting to push.
This is much more user-friendly and solves the pain of not being able to
add the tag somewhere in the history and thus effectively not being able
to push ever again because of the hook.

While implementing this change, it was necessary to change the prompt
package to start using special devices to make sure the hooks are
actually prompting user in the console. That means that /dev/tty is
opened on Unixes and \DEV\CON on Windows to make sure the user input is
read from the console.

It is also more or less required to have the SalsaFlow bootstrap
timestamp stored in the local configuration file to know what commits to
include in the check. There used to be some heuristics to get the right
commits in case the timestamp was not set, but these are now removed to
make the code much clearer. The timestamp is not that difficult to add
and the gain is significant.

Change-Id: db618e8032
Story-Id: SF-150