review post: Push in HEAD mode
From now on, review post pushes the current branch when in the HEAD
mode, i.e. when run as 'salsaflow review post'. This change makes the
behaviour more consistent.

The change was primarily introduces to fix GitHub when used for code
review to make sure the commit is pushed when the review issue is
ammended so that GitHub is able to create the link to the commit when
the review issue is edited, but in general this behaviour makes most
sense.

Change-Id: ed54793a64
Story-Id: SF-151