hooks/post-checkout: Skip check in review post
The post-checkout hook is run even when rewriting commits during 'review
post' and it may print the warning that there are Story-Id tags missing.
This is not needed in that case, so the hook was modified to just return
and not print anything in case the current branch is the temporary
branch.

Change-Id: f7ca6312e6
Story-Id: SF-156