review post: Fix critical bug
Not sure how this happened, but isStoryIdMissing was completely wrong.
In case there was a value not associated with the current issues
tracker, e.g. "unassigned", the commit was marked as missing the tag,
which is in this case obviously wrong. It is enough to check whether the
Story-Id tag is an empty string or not.

Story-Id: unassigned
Change-Id: 6bffeff828