review post: Always push when on branch tip
Make sure the current branch is in sync with the remote counterpart when
using any of the branch tip modes, i.e. the parent mode or the HEAD
mode. This makes it less probable for the developer to get confused
about the situation when setting Story-Id tag manually. In that case SF
was not pushing at all when no commit was amended, leading to the
developer having to push manually from time to time, which was easy to
forget.

SF is now pushing in any case when the current branch tip is not in
sync with the remote counterpart.

Change-Id: 75f0e01bd9
Story-Id: SF-158