github/codereview: Add review issues to milestones
Review issues, when created, are automatically added to the right review
milestones for the given release. The version number is the version
number as stored in the latest commit that is being associated with the
review issue being created.

This is actually not the optimal thing since the commit might be a bug
fix that is supposed to go to the release branch. In that case the
review milestone has to be changed manually. This is simply the best we
can do in guessing the review milestone.

Change-Id: c2218c72c8
Story-Id: SF-161