review stage: Fix rollback process
In case the final push is aborted, the rollback process is not correct
for release stage. For some reason it is not possible to recreate the
local release branch since that one already exists for unknown reason
since it is being deleted during the process.

It turns out that the failing push somehow creates the local branch
again. So to fix the rollback process, there is a check added and no
local release branch is created if it already exists.

Also fixes a couple of formatting and other issues found on the way.

Change-Id: fe5dbcdd72
Story-Id: SF-163