release start: Update the release workflow
Version suffixes are now incorporated into the release workflow, so we need to
bump the right versions with the right suffixes to the right branches. The
overall process looks like this:

  1) Create the release branch on top of develop.
  2) Bump version into the release branch.
  3) Bump version into the trunk branch.
  4) Push.

So compared to the original process, there is step 2, which was not necessary
when version suffixes were not being used.

Change-Id: 51c2b9948d
Story-Id: SF-160