release stage: Update the release workflow
The version handling has changed to incorporate version suffixes. This brings
changes for release stage as well since the version number has to be changed and
bumped for the staging branch when that one is created/reset.

So, the overall process looks like this:

  1) Reset the staging branch to the release branch.
  2) Delete the release branch.
  3) Bump version into the staging branch.
  4) Push.

Also, no tagging is happening during release stage any more, it is postponed
until release deploy, so this logic has been dropped from this command.

Change-Id: 4a41992986
Story-Id: SF-160