release deploy: Print more descriptive error
In case the release being closed is not releasable, there is no
descriptive error message for why that is happening.

This patch changes common.IssueTracker.Releasable into EnsureReleasable,
which returns an error in case the release is not releasable. All the
modules implementing this interface are turning a descriptive error
hint, listing which stories prevented the release step and why.

Change-Id: 0f4ab76106
Story-Id: SF-169