review post: Optionally list reviewed stories
The dialog for selecting a story to assign commits to has been largely
rewritten. It is possible to select a reviewed stories now as well, by
inserting 'r' to switch to a dialog listing the reviewed stories. The
same input can be used to switch back to the original dialog listing
stories that are being developed.

Also it's no longer '0' that represents no story. It has been changed to
'u' as in 'unassigned'.

Change-Id: 2f0c9388fe
Story-Id: SF-166