errs: Refactor the package heavily
Change the errs package to make more sense and to make it possible to
implement custom errors that work with this package. That basically
means that the package now uses Error interface instead of a struct, so
long as the custom error implements this interface, everything is cool.

Also got rid of bytes.Buffer in the log package, which didn't make much
sense and was actually supposed to be removed for a long time.

Change-Id: c683f71f72
Story-Id: SF-169