action: Add ActionChain struct
ActionChain is a chain of actions. You can push actions there, then call
Rollback and the actions will be rolled back in the FILO (stack) manner.
So calling ActionChain.Rollback() is pretty much the same as registering
multiple rollback actions using defer. The advantage of ActionChain is
that you can bundle the actions together and then return the whole
ActionChain, which itself implements action.Action interface.

Change-Id: a21d205ffa
Story-Id: SF-173