Improve git checks
Making sure that git branches are up to date and that the local tracking
branches exist should be more consistent now. There is a single
function, git.EnsureLocalTrackingBranch, that makes sure that the
associated remote branch exists, creates the local counterpart in case
it doesn't exist, and if it does exist, it makes sure the local branch
is up to date.

There are also some minor improvements in git checking scattered across
almost all commands.

Change-Id: b82edbedf3