review post: Show implemented dialog when appropriate
Show the implemented dialog iff the stories that were affected are
actually not marked as implemented yet. In other words, only show the
dialog when there are some stories that can be marked as implemented.

Also improved action.ActionChain a bit by actually returning an error
from Rollback() when there is an error in the bundled action rollback.

Change-Id: d493d5560a
Story-Id: SF-172