github/codereview: Create milestone when not found
The module is now more clever. It creates the review milestone when it
is not found when creating a review issue.

Even though the creation can happend on-demand like that, the release
init function still creates the milestone as well and that remains the
right time when the review milestone should be created. The on-demand
milestone creation is useful when starting with SalsaFlow on a project,
because no review milestone exists until release start is called.

Change-Id: 4eaed1551a
Story-Id: SF-176