releases: Fix ListNewTrunkCommits
This function was using the last tag to drop commits from the previous
release. This was not working any more because there can be a staged
release and another release already started at the same time. So we
actually want to drop the commits reachable from the staging branch,
not from the last tag.

Change-Id: aa6f061b77
Story-Id: SF-178