releases: Once again, fix ListNewTrunkCommits
Try to use the release branch before the staging branch in case the
release branch exists. This doesn't really matter now since the release
branch doesn't exist when this function is called, but in general we
want to make sure this function works no matter what.

In case there is no staging branch not release branch, limit the commits
by author date according to the SalsaFlow enabled timestamp. This
required also some refactoring in the git hooks.

Change-Id: f04e16fb4c
Story-Id: SF-178