jira: Fix/improve release start issue collection
Make sure that really all relevant issues are collected. The issue was
that when a subtask was encountered, the parent resource from that issue
was enqueued to be checked. But the JIRA API doesn't return complete
issue resources in the parent field. These resources are missing the
list of subtasks. In other words, issue.Parent.Fields.Subtasks is always
empty. So when a parent issue is to be enqueued, it is fetched from JIRA
to make sure the subtask list is there.

Also, when listing the issues that will be labeled automatically, the
list is split into tasks and subtasks and the reason why the issue was
added automatically is printed.

Change-Id: 86dca955bf
Story-Id: SF-179