Improve GitHub review issue parsing
This patch introduces a (hopefully) robust parser for GitHub review
issues. The package introduces, github/issues, will be used by the
daemon as well. The main point for the parser is that the parser
supports review blockers as well, which is just another checklist
following the commit checklist. Also it is possible to add user content
after a special separator string.

Check the string templates for particular review issue types to see what
the exact format is so that it is more clear what the parsers have to
handle.

The github/codereview package was modified to use this parser so that
there is no more stupid ad-hoc string handling.

Change-Id: d387ef842d
Story-Id: SF-185