github/codereview: Fix searching for review issues
The search was not using pagination at all, which could lead to the module not
finding some review issues. Also, pulling this into a separate module makes it
possible to be used by the daemon.

Actually the whole searching logic was wrong, because GitHub returns more issues
than just these matching the substring exactly, so it is necessary to go through
the issues and match them again locally to find the one exact substring match.

In case there are more issue matching, the first found is used.

Change-Id: 54c07cdcf6
Story-Id: SF-186