review post: Add -pick flag
This flag can be used to select the commits to be posted for code review
manually. After assigning the commits, the user is prompted to confirm
every commit (by default the commits are sent for code review). This
makes it possible to skip commits containing vendor dependencies and
other stuff that is not supposed to be reviewed.

Change-Id: 6194e57f84
Story-Id: SF-187