Add a global flag: -config
Now it is possible to specify the path to the global configuration file.
This is useful particularly when running `pkg upgrade` with sudo as
root. In this way you can specify your standard configuration file
without SalsaFlow trying to use the one for the root user, which is
usually missing since it is not usually needed.

Change-Id: dedcdcf4eb
Story-Id: SF-189