git: Change ShowCommits to ShowCommit
ShowCommits was actually only being called with one revision as the argument.
This change makes it possible to use git log instead of git show, which fixes a
bug. Using git show made the parsing logic go though the source code as well,
failing when encountering a line longer than 64 KB.

It would be possible to keep using show, but there is no reason for it.

Story-Id: unassigned
Change-Id: b7294f60fe