reviewboard: Run rbt interactively
Connect rbt to stdin, stdout and stderr of salsaflow itself so that it is
possible to send data to rbt in case it is needed interactively. This is easily
possible now since rbt is not executed concurrently for every commit, but
sequentially.

This fixes the bug with Review Board authentication needed. Before this change
the user simply got Username: EOF, but now it is possible to insert the
credentials interactively when needed.

Change-Id: 571a96abc7
Story-Id: SF-194