jira+pivotaltracker: Only assign unassigned stories
Up until now release start was assigning even these stories that were
already assigned to a release in case they modified trunk since the last
release. In practice this should not be happening, but when it happens,
we want to not assign the old stories to new release. We just print a
warning message.

Change-Id: 6b807952fd
Story-Id: SF-197