pivotaltracker: Revert to the origin workflow
The PT workflow has been changed to the one in GitFlow. The states are
represented now in the following way:

  * Being Implemented -> state:started
  * Implemented -> state:finished
  * Reviewed -> state:finished label:reviewed
  * Tested -> state: finished label:qa+

The skip labels (no qa, no review) are being taken into account as well.

Marking a story as implemented thus means to move it into Finished,
we are not adding any labels any more.

The other affected part is the release staging step. This has been changed to
reflect the changes in the workflow as well.

Change-Id: 83996d4266
Story-Id: SF-200