git: Rewrite FixCommitSources
FixCommitSources was calling CheckOrCreateTrackingBranch for the release
branch. That is, however, an issue during release stage. The local
branch is deleted and then the pre-push hook is run and since it uses
FixCommitSources, the local release branch is re-created, causing the
branch to exist even after 'release stage' is run.

This patch changes FixCommitSources not to use CheckOrCreateTracking
branch for the release branch. It only requires the local trunk branch
to exist.

Change-Id: b6b7c1be5f
Story-Id: SF-201