repo bootstrap: Push template sections into modules
The local config template now uses certain module-specific method to get
the local config section templates for all specified modules. This means
that there are no if-else conditions in the config template that mention
particular modules. This was necessary to make it possible really for
people to introduce 3rd party modules.

Change-Id: 83d6215c35
Story-Id: SF-211