review post: Pass implemented flag to the module
In case the user chooses to mark the affected stories as implemented, a
flag is passed into the code review module as "story_implemented" set to
true.

This change also affects the GitHub review module. In case the mentioned
flag is set, the review issue is labeled as implemented so that the
reviewer knows whether the review issue can be closed or not.

Change-Id: df82399e55
Story-Id: SF-215