pivotaltracker: Improve multiple repos support
Until now it was necessary to use a local git config to filter stories
affected by `story start` and the release commands. This is no longer
necessary, it is possible to commit the label into the local
configuration file. The key that needs to be set is
pivotal_tracker.component_label and it uses exact string matching, or
equality on labels to find the stories that should be shown/affected.

Change-Id: c65265aacf
Story-Id: SF-217