config: Refactor the package
Refactor the core package for reading and writing configuration files,
the main change being some changes in the config files schema and also
a change in encoding. We are moving from YAML to JSON since that allows
much easier deferred unmarshalling and pre-marshalling.

Some of the utility functions and methods are used by other packages that are
implemented in the following commits, namely config/loader.

Change-Id: e5620762f0
Story-Id: SF-220